fix(ohmypi): discover oversized model catalogs
