+++ crates/waku-core/src/lib.rs
@@ -52,6 +52,7 @@ pub mod workspace;
 pub mod worktree;
 
 mod fs_ext;
+mod pi_rpc;
 mod protocol;
 mod server;
